finalizando tela de locação e seus fluxos

+++ src/app/pages/private/locacao/locacao-insert-edit/locacao-insert-edit.component.ts
@@ -5,7 +5,7 @@ import {environment} from "../../../../../environments/environment";
 import {
   CLIENTE,
   CLIENTE_CONTATOS,
-  CLIENTE_ENDERECOS,
+  CLIENTE_ENDERECOS, CLIENTE_NOME,
   LOCACAO,
   LOCACAO_PRODUTOS,
   LOCACAO_SERVICOS
@@ -15,12 +15,11 @@ import {ContatoEntity} from "../../clientes/_entitys/contato.entity";
 import {ClientesEntity} from "../../clientes/clientes.entity";
 import {GenericService} from "../../../../services/generic-service/generic.service";
 import {CdkDynamicTable, CdkDynamicTableService} from "@datagrupo/dg-ng-util";
-import {ProdutosEntity} from "../../produtos/produtos.entity";
-import {ServicoEntity} from "../../servicos/servico.entity";
 import {ModalProdutoLocacaoComponent} from "../sub-components/modal-produto-locacao/modal-produto-locacao.component";
 import {ModalServicoLocacaoComponent} from "../sub-components/modal-servico-locacao/modal-servico-locacao.component";
 import Swal from "sweetalert2";
 import {AbstractInsertEdit2, InsertEditConfig2} from "@datagrupo/dg-crud";
+import {ActivatedRoute, Router} from "@angular/router";
 
 @Component({
   selector: 'app-locacao-insert-edit',
@@ -53,9 +52,12 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
   constructor(
     public config: InsertEditConfig2,
     public service: GenericService,
-    private CdkTable: CdkDynamicTableService
+    private CdkTable: CdkDynamicTableService,
+    private router: Router,
+    private route: ActivatedRoute
   ) {
     super(config)
+    this.rootConfig.subscribeEntityAfterSave = false;
 
     this.tableProtudos = this.CdkTable.create('request', {
       columns: [
@@ -69,24 +71,32 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
         context: LOCACAO_PRODUTOS
       },
       actions: {
-        edit: {name: 'Editar', action: val => this.modalProduto.open(val)},
+        edit: {name: 'Editar', action: val => this.modalProduto.open(val), dbClick: true},
         remove: {
           name: 'Remover', action: val => {
-            this.service.delete(LOCACAO_PRODUTOS + '/' + val.id).subscribe(
-              resp => {
-                this.tableProtudos.find();
-                this.atualizaTotais();
-                Swal.fire({
-                  icon: 'success',
-                  title: 'Produto removido'
-                }).then()
+            Swal.fire({
+              icon: 'question',
+              title: 'Remover produto desta locação',
+              showCancelButton: true,
+              cancelButtonText: 'Calcenar'
+            }).then(confirm => {
+              if (confirm.isConfirmed) {
+                this.service.delete(LOCACAO_PRODUTOS + '/' + val.id).subscribe(
+                  resp => {
+                    this.tableProtudos.find();
+                    this.atualizaTotais();
+                    Swal.fire({
+                      icon: 'success',
+                      title: 'Produto removido'
+                    }).then()
+                  }
+                )
               }
-            )
+            })
           }
         },
       }
     })
-
     this.tableServicos = this.CdkTable.create('request', {
       columns: [
         {name: 'nomeServico', headerName: 'Serviço', resource: val => val || '--'},
@@ -94,19 +104,28 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
       ],
       apiData: {path: environment.apiUrl, context: LOCACAO_SERVICOS},
       actions: {
-        edit: {name: 'Editar', action: val => this.modalServico.open(val)},
+        edit: {name: 'Editar', action: val => this.modalServico.open(val), dbClick: true},
         remove: {
           name: 'Remover', action: val => {
-            this.service.delete(LOCACAO_SERVICOS + '/' + val.id).subscribe(
-              resp => {
-                this.tableServicos.find();
-                this.atualizaTotais();
-                Swal.fire({
-                  icon: 'success',
-                  title: 'Serviço removido'
-                }).then()
+            Swal.fire({
+              icon: 'question',
+              title: 'Remover servoço desta locação',
+              showCancelButton: true,
+              cancelButtonText: 'Calcenar'
+            }).then(confirm => {
+              if (confirm.isConfirmed) {
+                this.service.delete(LOCACAO_SERVICOS + '/' + val.id).subscribe(
+                  resp => {
+                    this.tableServicos.find();
+                    this.atualizaTotais();
+                    Swal.fire({
+                      icon: 'success',
+                      title: 'Serviço removido'
+                    }).then()
+                  }
+                )
               }
-            )
+            })
           }
         },
       }
@@ -120,12 +139,25 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
 
     this.form.controls['cliente'].valueChanges.subscribe(
       resp => {
-        this.loadSelects(resp)
+        if (!this.entity?.id) {
+          this.loadSelects(resp)
+        }
+      }
+    )
+    this.form.controls['dataInicial'].valueChanges.subscribe(
+      resp => {
+        const date = new Date(resp)
+        const dateFinal = new Date()
+        dateFinal.setDate(date.getDate() + 30)
+
+        this.form.patchValue({
+          dataFinal: dateFinal.toLocaleString().split(', ')[0].split('/').reverse().join('-')
+        })
       }
     )
   }
 
-  loadSelects(cliente: number | string) {
+  loadSelects(cliente: number | string, reset: boolean = true) {
     this.service.get(CLIENTE_CONTATOS, {params: {cliente}}).subscribe(
       resp => {
         this.listContato = resp.data;
@@ -138,8 +170,10 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
       }
     )
 
-    this.form.controls['contato'].reset('');
-    this.form.controls['endereco'].reset('');
+    if (reset) {
+      this.form.controls['contato'].reset('');
+      this.form.controls['endereco'].reset('');
+    }
   }
 
   override ngOnInit(): void {
@@ -157,14 +191,13 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
     this.tableProtudos.controls.apiData.set({params: {locacao: this.entity.id}})
 
     if (!!this.entity?.cliente) {
-      this.listClientes = [
-        ...(this.entity.cliente ? [this.entity.cliente] : []),
-        ...this.listClientes
-      ]
+      this.listClientes.push(this.entity.cliente)
     }
 
+    this.form.controls['cliente'].disable();
+    this.loadSelects(this.entity.cliente?.id || '', false)
+
     if (this.entity.status == 'FINALIZADO') {
-      this.form.controls['cliente'].disable();
       this.form.controls['contato'].disable();
       this.form.controls['endereco'].disable();
       this.form.controls['dataInicial'].disable();
@@ -194,6 +227,12 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
     return true;
   }
 
+  override afterSaveEntity(entityData: any) {
+    this.router.navigate(['/user/locacoes/' + entityData.id]).then(() => {
+      window.location.reload();
+    })
+  }
+
   atualizaTotais() {
     if (!this.entity.id) return;
     this.service.get(LOCACAO + '/total/' + this.entity.id).subscribe(
@@ -208,5 +247,13 @@ export class LocacaoInsertEditComponent extends AbstractInsertEdit2<LocacaoEntit
 
   verifyEdit() {
     return ['ABERTO'].includes(this.entity.status || '')
+  }
+
+  findCliente(val:string) {
+    this.service.get(CLIENTE_NOME + val).subscribe(
+      resp => {
+        this.listClientes = resp.data;
+      }
+    )
   }
 }
